Clippy cleanup

+++ src/libxfuse/dir3_block.rs
@@ -30,6 +30,7 @@ use std::io::{BufRead, Seek, SeekFrom};
 use std::mem;
 use std::time::{Duration, UNIX_EPOCH};
 
+use super::S_IFMT;
 use super::da_btree::hashname;
 use super::definitions::*;
 use super::dinode::Dinode;
@@ -41,7 +42,7 @@ use super::utils::{get_file_type, FileKind};
 
 use byteorder::{BigEndian, ReadBytesExt};
 use fuser::{FileAttr, FileType};
-use libc::{c_int, ENOENT, S_IFMT};
+use libc::{c_int, ENOENT};
 
 pub const XFS_DIR2_DATA_FD_COUNT: usize = 3;
 
@@ -175,7 +176,7 @@ impl<R: BufRead + Seek> Dir3<R> for Dir2Block {
                 ),
                 crtime: UNIX_EPOCH,
                 kind,
-                perm: dinode.di_core.di_mode & (!(S_IFMT as u16)),
+                perm: dinode.di_core.di_mode & !S_IFMT,
                 nlink: dinode.di_core.di_nlink,
                 uid: dinode.di_core.di_uid,
                 gid: dinode.di_core.di_gid,
