+++ src/libxfuse/dir3_node.rs
@@ -29,6 +29,7 @@ use std::io::{BufRead, Seek, SeekFrom};
 use std::mem;
 use std::time::{Duration, UNIX_EPOCH};
 
+use super::S_IFMT;
 use super::bmbt_rec::BmbtRec;
 use super::da_btree::{hashname, XfsDa3Intnode};
 use super::definitions::*;
@@ -39,7 +40,7 @@ use super::utils::{get_file_type, FileKind};
 
 use byteorder::{BigEndian, ReadBytesExt};
 use fuser::{FileAttr, FileType};
-use libc::{c_int, ENOENT, S_IFMT};
+use libc::{c_int, ENOENT};
 
 #[derive(Debug)]
 pub struct Dir3FreeHdr {
@@ -206,7 +207,7 @@ impl<R: BufRead + Seek> Dir3<R> for Dir2Node {
             ),
             crtime: UNIX_EPOCH,
             kind,
-            perm: dinode.di_core.di_mode & (!(S_IFMT as u16)),
+            perm: dinode.di_core.di_mode & !S_IFMT,
             nlink: dinode.di_core.di_nlink,
             uid: dinode.di_core.di_uid,
             gid: dinode.di_core.di_gid,
